LogitSamplers

Documentation for LogitSamplers.

LogitSamplers.argmax_sampler

LogitSamplers.min_p_sampler

LogitSamplers.top_nσ_sampler

LogitSamplers.top_pk_sampler

LogitSamplers.argmax_sampler — Method

argmax_sampler(; device = identity)

Returns a function that samples most likely token.

LogitSamplers.min_p_sampler — Method

min_p_sampler(; pbase = 0.5f0, device = identity)

Returns a function that samples from the most probable tokens using the min-p strategy. See: https://arxiv.org/pdf/2407.01082

LogitSamplers.top_nσ_sampler — Method

top_nσ_sampler(; temperature = 1.0f0, n = 1.0f0, device = identity)

Returns a function that samples from the most probable tokens using the top-nσ strategy. See: https://arxiv.org/pdf/2411.07641

LogitSamplers.top_pk_sampler — Method

top_pk_sampler(; p = 0.5f0, k = 5, device = identity)

Returns a function that samples from the most probable tokens. A combination of the top-k and top-p sampling methods, where you can sample from the top tokens with cumulative probability p, with a max number of tokens k.